% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-methods.R
\name{Class_ForestTrain}
\alias{Class_ForestTrain}
\title{S3 class ForestTrain}
\description{
Class ForestTrain is the main class to contain models for binary classification forest/non-forest.
It includes the following elements:
}
\details{
In most cases objects of this class are generated by function \code{\link{train}}.
Then, classification of terrain images is made by \code{\link{classify}}.
}
\section{Slots}{
\tabular{ll}{
   Element \tab Description \cr
   call \tab the function call with which it was created \cr
   tp \tab the number of true positives obtained during training \cr
   fp \tab the number of false positives obtained during training \cr
   tn \tab the number of true negatives obtained during training \cr
   fn \tab the number of false negatives obtained during training \cr
}
}

